refact: move event types to model package
